Refactor: Update ip_address field type to Integer and add find_event endpoint for retrieving event details

+++ src/models/events.py
@@ -2,7 +2,7 @@ import uuid
 from typing import Annotated
 
 from pydantic import AnyHttpUrl, IPvAnyAddress, NonNegativeFloat, StringConstraints
-from sqlmodel import BigInteger, Field, String, Uuid
+from sqlmodel import Field, Integer, String, Uuid
 
 from .._types import Page
 from .base import BaseHyperModel
@@ -14,9 +14,7 @@ class Event(BaseHyperModel, table=True):
     req_id: uuid.UUID = Field(sa_type=Uuid, default_factory=uuid.uuid4)
     page: Page
     agent: Annotated[str, StringConstraints(min_length=10)]
-    ip_address: Annotated[IPvAnyAddress, Field(sa_type=BigInteger)]
-    # ip_address: str | None
+    ip_address: Annotated[IPvAnyAddress, Field(sa_type=Integer)]
     referrer: Annotated[AnyHttpUrl | None, Field(sa_type=String, nullable=True)]
-    # referrer: str | None = None
     session_id: str | None
     duration: NonNegativeFloat = 0
